requirements.txt
boto3==1.9.33
botocore==1.12.33
ccxt==1.17.427
celery==4.2.1
celery-redbeat==0.11.1
decorator==4.3.0
dj-database-url==0.5.0
Django==2.1.2
django-bmemcached==0.2.3
django-cache-memoize==0.1.2
django-cors-headers==2.4.0
django-debug-toolbar==1.10.1
django-extensions==2.1.3
django-filter==2.0.0
django-storages==1.7.1
django-unixtimestampfield==0.3.9
djangorestframework==3.7.7
gunicorn==19.9.0
jsonfield==2.0.2
numpy==1.16.4
psycopg2-binary==2.7.6.1
python-binary-memcached==0.28.0
python-dateutil==2.7.4
pytz==2018.6
PyYAML==3.13
redis==2.10.6
requests==2.20.0
requests-oauthlib==1.0.0
schedule==0.5.0
simplejson==3.16.0
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
traitlets==4.3.2
typing==3.6.6
uhashring==1.1
urllib3==1.24
waitress==1.1.0
whitenoise==4.1
yarl==1.2.6


## FANCY EXTRAS
artemis-ml==2.0.0
cloudpickle==0.6.1
deap==1.2.2
empyrical==0.5.0
pandas==0.23.4
pathos==0.2.3
pyfolio==0.9.0
TA-Lib==0.4.17
tqdm==4.28.1


## ACTUALLY NEEDED??
# aiodns==1.1.1
# aiohttp==3.4.4
# amqp==2.3.2
# appnope==0.1.0
# async-timeout==3.0.1
# attrs==18.2.0
# backcall==0.1.0
# billiard==3.5.0.4
# boto==2.49.0
# certifi==2018.10.15
# chardet==3.0.4
# docutils==0.14
# idna==2.7
# idna-ssl==1.1.0
# ipython==7.0.1
# ipython-genutils==0.2.0
# jedi==0.13.1
# jmespath==0.9.3
# kombu==4.2.1
# multidict==4.4.2
# oauthlib==2.1.0
# parso==0.3.1
# pexpect==4.6.0
# pickleshare==0.7.5
# prompt-toolkit==2.0.6
# Pygments==2.2.0
# ptyprocess==0.6.0
# pycares==2.3.0
# s3transfer==0.1.13
# simplegeneric==0.8.1
# toolz==0.9.0
# vine==1.1.4
# wcwidth==0.1.7


## ONLY FOR LOCAL
# networkx==2.2
# graphviz==0.10.1
# xlsxwriter==1.1.7
